Move stashy to java/stashy
