+++ src/main/java/stashy/Stashy.java
@@ -1,11 +1,11 @@
-package main.stashy;
+package stashy;
 
-import main.stashy.commands.Command;
-import main.stashy.data.exception.StashyException;
-import main.stashy.data.task.TaskList;
-import main.stashy.parser.Parser;
-import main.stashy.storage.Storage;
-import main.stashy.ui.Ui;
+import stashy.commands.Command;
+import stashy.data.exception.StashyException;
+import stashy.data.task.TaskList;
+import stashy.parser.Parser;
+import stashy.storage.Storage;
+import stashy.ui.Ui;
 
 /**
  * Represents a chatbot that can do many things, just like
