+++ src/main/java/stashy/data/task/Deadline.java
@@ -1,4 +1,4 @@
-package main.stashy.data.task;
+package stashy.data.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
