+++ src/main/java/stashy/data/task/ToDo.java
@@ -1,4 +1,4 @@
-package main.stashy.data.task;
+package stashy.data.task;
 
 /**
  * One of the task types, specifically those with
